refactor: move theme, color and theme function to styles
