+++ sepal_ui/frontend/styles.py
@@ -1,3 +1,6 @@
+from pathlib import Path
+from configparser import ConfigParser
+from types import SimpleNamespace
 from traitlets import Unicode
 from IPython.display import display
 import ipyvuetify as v
@@ -24,6 +27,51 @@ v.theme.themes.light.bg_color = "#fff"
 # set a specific color for menus
 v.theme.themes.dark.menu = "#424242"
 v.theme.themes.light.menu = "#fff"
+
+config_file = Path.home() / ".sepal-ui-config"
+"Pathlib.Path: the configuration file generated by sepal-ui based application to save parameters as language or theme"
+
+
+def get_theme(config_file):
+    """
+    get the theme from the config file (default to dark)
+
+    Return:
+        (str): the theme to use
+    """
+
+    # init theme
+    theme = "dark"
+
+    # read the config file if existing
+    if config_file.is_file():
+        config = ConfigParser()
+        config.read(config_file)
+        theme = config.get("sepal-ui", "theme", fallback="dark")
+
+    # set vuetify theme
+    v.theme.dark = True if theme == "dark" else False
+
+    return theme
+
+
+theme = getattr(v.theme.themes, get_theme(config_file))
+"traitlets: the theme used in sepal"
+
+color = SimpleNamespace(
+    main=theme.main,
+    darker=theme.darker,
+    bg=theme.bg_color,
+    primary=theme.primary,
+    accent=theme.accent,
+    secondary=theme.secondary,
+    success=theme.success,
+    info=theme.info,
+    warning=theme.warning,
+    error=theme.error,
+    menu=theme.menu,
+)
+'SimpleNamespace: the colors of sepal. members are in the following list: "main, darker, bg, primary, accent, secondary, success, info, warning, error, menu". They will render according to the selected theme.'
 
 
 class Styles(v.VuetifyTemplate):
@@ -92,4 +140,25 @@ ICON_TYPES = {
     ".geojson": _vector,
     "DEFAULT": _other,
     "PARENT": _parent,
+}
+
+
+# Set default styles to layers
+
+layer_style = {  # default styling of the layer
+    "stroke": True,
+    "color": color.primary,
+    "weight": 2,
+    "opacity": 1,
+    "fill": True,
+    "fillOpacity": 0,
+}
+
+layer_hover_style = {  # current alert styling
+    "stroke": True,
+    "color": color.primary,
+    "weight": 5,  # same weight as hover to hide the general layer
+    "opacity": 1,
+    "fill": True,  # to avoid unwanted click on the underlying alert
+    "fillOpacity": 0,
 }
